// src/generated/enums.ts
export const EBetaBranchFlags = {
  k_EBetaBranch_None: 0,
  k_EBetaBranch_Default: 1,
  k_EBetaBranch_Available: 2,
  k_EBetaBranch_Private: 4,
  k_EBetaBranch_Selected: 8,
  k_EBetaBranch_Installed: 16,
} as const;
// ...
export const EGamepadTextInputMode = {
  k_EGamepadTextInputModeNormal: 0,
  k_EGamepadTextInputModePassword: 1,
} as const;

export const EGamepadTextInputLineMode = {
  k_EGamepadTextInputLineModeSingleLine: 0,
  k_EGamepadTextInputLineModeMultipleLines: 1,
} as const;
// ...
export const ESteamHardwareType = {
  k_ESteamHardwareTypeNone: 0,
  k_ESteamHardwareTypeSteamDeck: 1,
  k_ESteamHardwareTypeSteamMachine: 2,
  k_ESteamHardwareTypeSteamFrame: 3,
} as const;